Move MediawikiTextWrapper to production config
The MediawikiTextWrapper webpack plugin wraps the JavaScript generated by
Webpack in WikiText that can be pasted in CentralNotice. Since commit
88ee3d48f2c0810b73e7f67eaf71cbca20cc9b1b it suppresses the JavaScript
files that will be written, to keep the build directory clean.

This commit moves the MediawikiTextWrapper to the webpack production
configuration setup, because the development environment uses the
JavaScript files generated by Webpack.
