+++ webpack.common.js
@@ -2,18 +2,12 @@ const path = require( 'path' );
 const fs = require( 'fs' );
 const toml = require( 'toml' );
 const webpack = require( 'webpack' );
-const MediaWikiTextWrapper = require( './webpack/mediawiki_text_wrapper' );
-const CampaignConfig = require( './webpack/campaign_config' );
 const WrapperPlugin = require( 'wrapper-webpack-plugin' );
 
+const CampaignConfig = require( './webpack/campaign_config' );
 const campaigns = new CampaignConfig( toml.parse( fs.readFileSync( 'campaign_info.toml', 'utf8' ) ) );
 
-function readWrapperTemplate( name ) {
-	return fs.readFileSync( './webpack/wikitext_templates/' + name + '.hbs', 'utf8' );
-}
-
 module.exports = {
-	devtool: 'sourcemap',
 	entry: campaigns.getEntryPoints(),
 	output: {
 		filename: '[name].js',
@@ -67,15 +61,6 @@ module.exports = {
 	plugins: [
 		new webpack.ProvidePlugin( {
 			jQuery: 'jquery'
-		} ),
-		new MediaWikiTextWrapper( {
-			templates: campaigns.getWrapperTemplates( readWrapperTemplate ),
-			context: {
-				bannerValuesJS: '{{MediaWiki:WMDE_FR2017/Resources/BannerValues.js}}',
-				bannerValues: '{{MediaWiki:WMDE_Fundraising/Campaign_Parameters_2019}}'
-			},
-			filePattern: '{B,WMDE}*.js',
-			campaignConfig: campaigns.getConfigForPages()
 		} ),
 		new WrapperPlugin( {
 			test: /B\d{2}WPDE.*.js$/,
