+++ webpack.config.js
@@ -5,6 +5,7 @@ const CommonConfig = require( './webpack.common.js' );
 const webpack = require( 'webpack' );
 
 module.exports = Merge( CommonConfig, {
+	devtool: 'source-map',
 	mode: 'development',
 	entry: {
 		loader: './webpack/loader.js'
